Operators #time 5m

+++ control-flow/solution/control_input.py
@@ -3,7 +3,7 @@ AGE_ADULT = 18
 
 age = input('Podaj wiek: ')
 
-if int(age) >= AGE_ADULT:
+if float(age) >= AGE_ADULT:
     print('Adult')
 else:
     print('Young')
